requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
mysql-connector-python==8.0.26
bcrypt==3.2.0
Faker==8.12.1
WTForms==2.3.3
email-validator==1.1.3 